+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.4]
+
+### Added
+
+- Test additional resources in unit tests also
+
+### Fixed
+
+- Fixed iam integration when `module_enabled = false`
+
 ## [0.0.3]
 
 ### Added
@@ -35,7 +45,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Implementation
 
-[unreleased]: https://github.com/mineiros-io/terraform-google-artifact-registry-repository/compare/v0.0.3...HEAD
+[unreleased]: https://github.com/mineiros-io/terraform-google-artifact-registry-repository/compare/v0.0.4...HEAD
+[0.0.4]: https://github.com/mineiros-io/terraform-google-artifact-registry-repository/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/mineiros-io/terraform-google-artifact-registry-repository/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/mineiros-io/terraform-google-artifact-registry-repository/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/mineiros-io/terraform-google-artifact-registry-repository/releases/tag/v0.0.1